Reworked main file
Major rework of main file and library files
